distro qcom-base.inc: enable EFI distro feature, select ipk as packageclass
EFI is a combined machine/distro feature, so it needs the DISTRO to
enable it. This has no boot impact for non-EFI machines, the main impact
is that optional EFI tools are now actually built.

The package class change ensures that we keep the current default, so
upstream changes won't affect this DISTRO.

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>