qcom distro: rename, add X11 based distro config
This adds `x11` and `vulkan` to DISTRO_FEATURES, enabling graphical
environments like XFCE. It should be safe to enable both `x11` and
`wayland`, other DISTROs have been doing this for a while.

Since the DISTRO is no longer wayland specific rename it and its
overrides. The meta-qcom-distro isn't using any of the `:qcom-wayland` overrides
currently.

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>