qcom-base.inc: enable tarballs as additional image format
Enable `tar.gz` to be able to populate external media like microSD cards
a bit easier. OE-core has `tar.gz` as its default, but BSP layers tend
to overwrite that, so DISTROs need to use `:append` to ensure the
desired types get generated.

Gzip has been picked over xz because of its speed, both in compression
as in decompression. With `pigz` installed it scales linearly with the
number of CPU cores, removing it as the bottleneck.

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>