packagegroups: Define QCOM userspace utilities packagegroup
Define a package group with common utilities required for QCOM images.
Previously, these utilities were added to packagegroup-qcom via a
bbappend. However, some of the utilities, like libinput, force the
PACKAGE_ARCH to be updated, which can cause confusion for users.
Dropping the append and adding these utilities in a new packagegroup
would be a much cleaner implementation.

Signed-off-by: Viswanath Kraleti <quic_vkraleti@quicinc.com>