layer: add dependency on openembedded-layer
Add explicit layer dependency on openembedded-layer to satisfy
dependency required for adb support.

While we could change the image_adbd class to only depend on
android-tools-adbd when openembedded-layer is available, it is better to
assume it will always be for better user experience.

Since we also expect to extend our distro images to also include
additional recipes from the same layer, better to just have it as an
explicit dependency.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>