qcom-base: switch to glvnd implementation
GLVND is a mechanism to provide several OpenGL / GL ES / EGL
implementations, allowing one to use the one which works on a particular
platform. This follow the ideas of ICD (installable client drivers) as
used by Vulkan and OpenCL.

Switch the distro to using GLVND in order to allow co-installing Mesa
with other OpenGL implementations.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>