Github workflows: build-yocto.yml: restrict build to 3 machines
The CI only needs to test if the DISTROs work, not the BSPs itself,
so restrict the build to the following 3 machines:
 * qcom-armv7a
 * qcom-armv8a
 * rb3gen2-core-kit

Restrict the KVM build to qcom-armv8a, it is not supported on armv7a
machines.

As a consequence: drop the warmup task, it isn't needed with the much
reduced scope.

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>